+++ template/lugia/class/view/modals/DetailModal.js
@@ -119,6 +119,11 @@ class DetailModal extends PureComponent {
 					type: 'TableListByName',
 					config: formServiceConfig,
 					disabled: isOpenFormModal.type !== 'add',
+					dropDownStyle: {
+						maxHeight: '400px',
+					},
+					dropdownMatchSelectWidth: false,
+					getPopupContainer: () => document.getElementById('$PageNameDetailModal'),
 					allowClear: true,
 					onChange(value, option) {
 						if (value) {
@@ -145,6 +150,11 @@ class DetailModal extends PureComponent {
 					config: formServiceConfig,
 					allowClear: true,
 					disabled: true,
+					dropDownStyle: {
+						maxHeight: '400px',
+					},
+					dropdownMatchSelectWidth: false,
+					getPopupContainer: () => document.getElementById('$PageNameDetailModal'),
 					onChange(value) {},
 				},
 			},
@@ -163,6 +173,11 @@ class DetailModal extends PureComponent {
 					allowClear: true,
 					placeholder: '请选择评级类型',
 					disabled: showdetails,
+					dropDownStyle: {
+						maxHeight: '400px',
+					},
+					dropdownMatchSelectWidth: false,
+					getPopupContainer: () => document.getElementById('$PageNameDetailModal'),
 					onChange(value) {
 						let parentDicCode = '';
 						if (value === '10090001') {
@@ -200,6 +215,7 @@ class DetailModal extends PureComponent {
 					disabled: showdetails,
 					placeholder: '请选择日期',
 					initialValue: moment(),
+					getCalendarContainer: () => document.getElementById('$PageNameDetailModal'),
 					onChange(value) {},
 				},
 			},
@@ -219,6 +235,11 @@ class DetailModal extends PureComponent {
 					placeholder: '请选择信用评级',
 					disabled: showdetails,
 					onChange(value) {},
+					dropDownStyle: {
+						maxHeight: '400px',
+					},
+					dropdownMatchSelectWidth: false,
+					getPopupContainer: () => document.getElementById('$PageNameDetailModal'),
 				},
 			},
 			{
@@ -283,13 +304,15 @@ class DetailModal extends PureComponent {
 			},
 		];
 		return (
-			<NeatForm
-				refs={(ref) => (this.createProduct = ref)}
-				labelSize='100px'
-				lineOf={2}
-				formItem={formItems}
-				viewing={isOpenFormModal.type === 'detail' ? true : false}
-			/>
+			<div id='$PageNameDetailModal'>
+				<NeatForm
+					refs={(ref) => (this.createProduct = ref)}
+					labelSize='100px'
+					lineOf={2}
+					formItem={formItems}
+					viewing={isOpenFormModal.type === 'detail' ? true : false}
+				/>
+			</div>
 		);
 	}
 }