+++ template/lugia/hooks/view/modals/DetailModal.js
@@ -20,14 +20,13 @@ const DetailModal = (props, ref) => {
 		//  暴露给父组件的方法
 		handleSubmit: handleSubmit,
 	}));
-	const { isOpenFormModal, asyncHttpGetCreditRateDropdownList, creditRateDropdownList } = props;
+	const { isOpenFormModal, creditRateDropdownList, projectRowed, asyncHttpGetListData, changeSync } = props;
 	const showdetails = isOpenFormModal.type === 'detail';
 	const [rateCompName, setRateCompName] = useState(undefined);
 
 	let _this = this;
 
 	useEffect(() => {
-		const { isOpenFormModal, projectRowed } = props;
 		props.onRef(this);
 		//表单初始化
 		if (projectRowed.rateType != null) {
@@ -60,7 +59,6 @@ const DetailModal = (props, ref) => {
 
 	//点击确定进行增加修改操作
 	const handleSubmit = () => {
-		const { asyncHttpGetListData, projectRowed, changeSync, isOpenFormModal } = props;
 		// e.preventDefault();
 		createProduct.onValidate(
 			(values) => {
@@ -118,6 +116,11 @@ const DetailModal = (props, ref) => {
 				config: formServiceConfig,
 				disabled: isOpenFormModal.type !== 'add',
 				allowClear: true,
+				dropDownStyle: {
+					maxHeight: '400px',
+				},
+				dropdownMatchSelectWidth: false,
+				getPopupContainer: () => document.getElementById('$PageNameDetailModal'),
 				onChange(value, option) {
 					if (value) {
 						createProduct.setValues({
@@ -143,6 +146,11 @@ const DetailModal = (props, ref) => {
 				config: formServiceConfig,
 				allowClear: true,
 				disabled: true,
+				dropDownStyle: {
+					maxHeight: '400px',
+				},
+				dropdownMatchSelectWidth: false,
+				getPopupContainer: () => document.getElementById('$PageNameDetailModal'),
 				onChange(value) {},
 			},
 		},
@@ -161,6 +169,11 @@ const DetailModal = (props, ref) => {
 				allowClear: true,
 				placeholder: '请选择评级类型',
 				disabled: showdetails,
+				dropDownStyle: {
+					maxHeight: '400px',
+				},
+				dropdownMatchSelectWidth: false,
+				getPopupContainer: () => document.getElementById('$PageNameDetailModal'),
 				onChange(value) {
 					let parentDicCode = '';
 					if (value === '10090001') {
@@ -196,6 +209,7 @@ const DetailModal = (props, ref) => {
 				disabled: showdetails,
 				placeholder: '请选择日期',
 				initialValue: moment(),
+				getCalendarContainer: () => document.getElementById('$PageNameDetailModal'),
 				onChange(value) {},
 			},
 		},
@@ -277,13 +291,15 @@ const DetailModal = (props, ref) => {
 		},
 	];
 	return (
-		<NeatForm
-			refs={(ref) => (createProduct = ref)}
-			labelSize='100px'
-			lineOf={2}
-			formItem={formItems}
-			viewing={isOpenFormModal.type === 'detail' ? true : false}
-		/>
+		<div id='$PageNameDetailModal'>
+			<NeatForm
+				refs={(ref) => (createProduct = ref)}
+				labelSize='100px'
+				lineOf={2}
+				formItem={formItems}
+				viewing={isOpenFormModal.type === 'detail' ? true : false}
+			/>
+		</div>
 	);
 };
 export default forwardRef(DetailModal);