+++ CHANGELOG.md
@@ -19,3 +19,4 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 -   updated template, version `0.3.1` (Thu Feb 10 17:26:00 2022 +0800)
 -   updated template, version `0.3.4` (Mon May 16 09:26:00 2022 +0800)
 -   updated react hooks template, version `0.3.6` (Mon Jun 06 15:25:00 2022 +0800)
+-   updated react hooks template, version `0.3.7` (Mon Jun 09 17:29:00 2022 +0800)